+++ config/cors.php
@@ -21,6 +21,7 @@ return [
         'https://fitlifemilano.it',
         'http://127.0.0.1:8000',
         'http://localhost:8000',
+        'http://fitlife.test:8000',
     ],
 
     'allowed_origins_patterns' => [],